app/requirements.txt
flask==0.12.2
requests==2.18.0
pyyaml==3.13
jinja2==2.10
django==2.0.1
cryptography==2.1.4
lxml==4.6.2
paramiko==2.7.1
Pillow==8.1.0
